Oraculo/requirements.txt
# Framework Web
streamlit==1.38.0

# LangChain
langchain==0.3.0
langchain-community==0.3.0
langchain-openai==0.2.0

# Processamento de Documentos - Existentes
pypdf==5.0.0
beautifulsoup4==4.12.3
unstructured==0.15.13
youtube_transcript_api>=0.6.2

# Processamento de Documentos - Novos
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23

# OCR
pytesseract==0.3.10
Pillow==10.4.0

# Utilidades
python-dotenv==1.0.1
fake_useragent==1.5.1
numpy
